requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.5
langchain-community>=0.3.0
langchain-core>=0.3.0

# Neo4j
neo4j>=5.11.0

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
numpy>=1.24.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.5.0
